requirements.txt
pyyaml
requests
torch
numpy
Pillow
google-generativeai